+++ tests/page-templates/00-basic/00-onload.js
@@ -48,6 +48,9 @@ describe("e2e/00-basic/00-onload", function() {
 
 		assert.isNumber(tf.lastBeacon()["rt.bstart"], "rt.bstart");
 		assert.isNumber(tf.lastBeacon()["rt.end"], "rt.end");
+		assert.isString(tf.lastBeacon()["rt.si"], "rt.si");
+		assert.isNumber(tf.lastBeacon()["rt.sl"], "rt.sl");
+		assert.isNumber(tf.lastBeacon()["rt.ss"], "rt.ss");
 		assert.isString(tf.lastBeacon()["rt.start"], "rt.start");
 
 		// optional
@@ -84,7 +87,7 @@ describe("e2e/00-basic/00-onload", function() {
 		assert.isString(tf.lastBeacon().pid, "pid");
 	});
 
-	it("Should not have set nocookie=1", function() {
+	it("Should have not set nocookie", function() {
 		assert.isUndefined(tf.lastBeacon().nocookie);
 	});
 });