+++ tests/page-templates/14-errors/23-autorun-false.js
@@ -57,6 +57,11 @@ describe("e2e/14-errors/23-autorun-false", function() {
 		assert.isDefined(b["rt.end"]);
 	});
 
+	it("Should have rt.sl = 0", function() {
+		var b = tf.lastBeacon();
+		assert.equal(b["rt.sl"], 0);
+	});
+
 	it("Should have rt.tstart", function() {
 		var b = tf.lastBeacon();
 		assert.isDefined(b["rt.tstart"]);