+++ tests/page-templates/14-errors/24-autorun-false-multiple-then-load.js
@@ -96,4 +96,14 @@ describe("e2e/14-errors/24-autorun-false-multiple-then-load", function() {
 			return this.skip();
 		}
 	});
+
+	it("Should have rt.sl = 0 on the first beacon", function() {
+		var b = tf.beacons[0];
+		assert.equal(b["rt.sl"], 0);
+	});
+
+	it("Should have rt.sl = 1 on the second beacon", function() {
+		var b = tf.beacons[1];
+		assert.equal(b["rt.sl"], 1);
+	});
 });