+++ tests/unit/01-basic.js
@@ -16,6 +16,10 @@ describe("BOOMR exports", function() {
 		assert.isString(BOOMR.version);
 	});
 
+	it("Should have an existing BOOMR.session Object", function() {
+		assert.isObject(BOOMR.session);
+	});
+
 	it("Should have an existing BOOMR.init() Function", function() {
 		assert.isFunction(BOOMR.init);
 	});